Migrate Printer and MessageStream to submodule
